vector problem fix

+++ NMR_2015revamp/coords_3d.cpp
@@ -6,25 +6,25 @@
  */
 #include "coords_3d.h"
 
-float coords_3d::getX() {
+double coords_3d::getX() {
   return x;
 }
-void coords_3d::setX(float x_given) {
+void coords_3d::setX(double x_given) {
   x = x_given;
 }
-float coords_3d::getY() {
+double coords_3d::getY() {
   return y;
 }
-void coords_3d::setY(float y_param) {
+void coords_3d::setY(double y_param) {
   y = y_param;
 }
-float coords_3d::getZ() {
+double coords_3d::getZ() {
   return z;
 }
-void coords_3d::setZ(float z_param) {
+void coords_3d::setZ(double z_param) {
   z = z_param;
 }
-coords_3d::coords_3d(float x_given, float y_given, float z_given){
+coords_3d::coords_3d(double x_given, double y_given, double z_given){
   x=x_given;
   y=y_given;
   z=z_given;
