+++ NMR_2015revamp/coords_3d.h
@@ -9,15 +9,15 @@
 #define COORDS_3D_H_
 #include <string>
 class coords_3d{
-  float x,y,z;
+  double x,y,z;
  public:
-  float getX();
-  void setX(float x);
-  float getY();
-  void setY(float y);
-  float getZ();
-  void setZ(float z);
-  coords_3d(float,float,float);
+  double getX();
+  void setX(double x);
+  double getY();
+  void setY(double y);
+  double getZ();
+  void setZ(double z);
+  coords_3d(double,double,double);
   coords_3d();
 };
