+++ NMR_2015revamp/vector_3d.h
@@ -11,23 +11,20 @@
 #include <cmath>
 
 class vector_3d {
-      float x_coord,y_coord,z_coord;
-	public:
-      float get_x();
-      float get_y();
-      float get_z();
-      void set_x(float);
-      void set_y(float);
-      void set_z(float);
-      float get_dot_product(vector_3d *);
-      float get_magnitude();
-      float get_angle_from(vector_3d *);
-      vector_3d(float, float,float);
-      vector_3d();
-      vector_3d(coords_3d,coords_3d);
+  double x_coord, y_coord, z_coord;
+ public:
+  double get_x();
+  double get_y();
+  double get_z();
+  void set_x(double);
+  void set_y(double);
+  void set_z(double);
+  double get_dot_product(vector_3d *);
+  double get_magnitude();
+  double get_angle_from(vector_3d *);
+  vector_3d(double, double, double);
+  vector_3d();
+  vector_3d(coords_3d, coords_3d);
 };
-
-
-
 
 #endif /* 3DVECTOR_H_ */
